refactor: Player, Dealer 공통으로 사용할 수 있는 부분 수정
