+++ src/main/java/blackjack/domain/user/Players.java
@@ -1,6 +1,5 @@
 package blackjack.domain.user;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
