+++ src/main/java/blackjack/domain/user/User.java
@@ -7,8 +7,10 @@ import blackjack.domain.card.CardDeck;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 public abstract class User implements GameRule {
+    private static final String NULL_ERR_MSG = "%s이(가) 없습니다.";
     protected static int INITIAL_CARD_SIZE = 2;
     protected static int ADDITIONAL_CARD_SIZE = 1;
 
@@ -24,9 +26,27 @@ public abstract class User implements GameRule {
         return name;
     }
 
-    abstract public void drawCard(CardDeck deck);
-
     public List<Card> getCards() {
         return cards;
+    }
+
+    public void drawCard(CardDeck deck) {
+        Objects.requireNonNull(deck, String.format(NULL_ERR_MSG, "카드"));
+        int receivableCardSize = getReceivableCardSize();
+        for (int i = 0; i < receivableCardSize; i++) {
+            Card card = deck.getCard();
+            getCards().add(card);
+        }
+    }
+
+    @Override
+    public int getReceivableCardSize() {
+        if (getCards().size() == 0) {
+            return INITIAL_CARD_SIZE;
+        }
+        if (receivable()) {
+            return ADDITIONAL_CARD_SIZE;
+        }
+        return 0;
     }
 }
